pyproject.toml
[project]
name = "review-master"
version = "0.1.0"
description = "Multi-tenant review management platform — Superadmin module"
requires-python = ">=3.12"
readme = "README.md"
dependencies = [
  "django==6.0.2",
  "djangorestframework==3.17.1",
  "django-vite==3.1.0",
  "django-redis==5.4.0",
  "django-environ==0.11.2",
  "psycopg[binary]==3.2.3",
  "python-json-logger==3.2.0",
  "drf-spectacular==0.28.0",
]

[dependency-groups]
dev = [
  "pytest==8.3.3",
  "pytest-django==4.9.0",
  "pytest-cov==5.0.0",
  "factory-boy==3.3.3",
  "pre-commit==4.0.1",
  "ruff==0.15.11",
  "mypy==1.13.0",
  "django-stubs[compatible-mypy]==5.1.1",
  "djangorestframework-stubs==3.15.1",
  "bandit[toml]==1.8.0",
  "django-debug-toolbar==4.4.6",
]

[tool.ruff]
target-version = "py312"
line-length = 100
extend-exclude = ["migrations", "static", "media", "node_modules"]

[tool.ruff.lint]
select = ["E", "W", "F", "I", "UP", "B", "C4", "DJ", "SIM", "RUF", "S", "N", "T20", "PT", "TID"]
ignore = ["E501", "S101"]

[tool.ruff.lint.per-file-ignores]
"**/tests/**" = ["S101", "S106", "ARG"]
"**/migrations/**" = ["E501", "N806"]
"**/settings/**" = ["F403", "F405"]
"apps/*/apps.py" = ["F401"]

[tool.ruff.lint.isort]
known-first-party = ["apps", "config"]
section-order = ["future", "standard-library", "third-party", "first-party", "local-folder"]

[tool.ruff.format]
quote-style = "double"
indent-style = "space"

[tool.mypy]
python_version = "3.12"
strict = true
plugins = ["mypy_django_plugin.main", "mypy_drf_plugin.main"]
exclude = ["migrations/", "venv/", ".venv/"]

[[tool.mypy.overrides]]
module = ["environ.*"]
ignore_missing_imports = true

[[tool.mypy.overrides]]
module = ["factory.*"]
ignore_missing_imports = true

[[tool.mypy.overrides]]
module = ["apps.*.tests.*", "apps.*.tests"]
ignore_errors = true

[tool.django-stubs]
django_settings_module = "config.settings.test"

[tool.pytest.ini_options]
DJANGO_SETTINGS_MODULE = "config.settings.test"
python_files = ["test_*.py"]
addopts = "--strict-markers --tb=short -q"

[tool.coverage.run]
source = ["apps"]
omit = ["*/migrations/*", "*/tests/*"]

[tool.coverage.report]
fail_under = 85

[tool.bandit]
exclude_dirs = ["tests", "migrations", ".venv"]
